pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deploy-service"
version = "0.1.0"
description = "Deploy one TrueFoundry service: multiple stdio MCP servers in one server (TBXark mcp-proxy)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.24.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
]

[project.scripts]
deploy-service = "deploy_service.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
